ReplicateWrapper/requirements.txt
fastapi
uvicorn 
python-dotenv

replicate
